+++ Tests/Test-MongoFiles.ps1
@@ -22,59 +22,57 @@ $total = $collection.Count()
 "$total documents"
 
 "EQ NE"
-$EQReadme = Test-Query (query Name Readme.txt)
-$NEReadme = Test-Query (query Name -NE Readme.txt)
+$EQReadme = Test-Query (New-MdbcQuery Name Readme.txt)
+$NEReadme = Test-Query (New-MdbcQuery Name -NE Readme.txt)
 if ($total -ne $EQReadme + $NEReadme) { throw "$total -ne $EQReadme + $NEReadme" }
 
 "IEQ INE"
-$n1 = Test-Query (query Name -IEQ README.TXT)
+$n1 = Test-Query (New-MdbcQuery Name -IEQ README.TXT)
 if ($n1 -lt $EQReadme) { throw }
-$n2 = Test-Query (query Name -INE README.TXT)
+$n2 = Test-Query (New-MdbcQuery Name -INE README.TXT)
 if ($n2 -ne $total - $n1) { throw }
 
 "EQ GT LT"
-$n1 = Test-Query (query LastWriteTime $time1)
-$n2 = Test-Query (query LastWriteTime -GT $time1)
-$n3 = Test-Query (query LastWriteTime -LT $time1)
+$n1 = Test-Query (New-MdbcQuery LastWriteTime $time1)
+$n2 = Test-Query (New-MdbcQuery LastWriteTime -GT $time1)
+$n3 = Test-Query (New-MdbcQuery LastWriteTime -LT $time1)
 if ($total -ne $n1 + $n2 + $n3) { throw }
 
 "GE LE"
-$n2 = Test-Query (query LastWriteTime -GE $time1)
-$n3 = Test-Query (query LastWriteTime -LE $time1)
+$n2 = Test-Query (New-MdbcQuery LastWriteTime -GE $time1)
+$n3 = Test-Query (New-MdbcQuery LastWriteTime -LE $time1)
 if ($total -ne -$n1 + $n2 + $n3) { throw }
 
 "And"
-$n1 = Test-Query (query (query Name Readme.txt), (query LastWriteTime -GT $time1))
-$n2 = Test-Query (query (query Name Readme.txt), (query LastWriteTime -LT $time1))
+$n1 = Test-Query (New-MdbcQuery (New-MdbcQuery Name Readme.txt), (New-MdbcQuery LastWriteTime -GT $time1))
+$n2 = Test-Query (New-MdbcQuery (New-MdbcQuery Name Readme.txt), (New-MdbcQuery LastWriteTime -LT $time1))
 if ($EQReadme -ne $n1 + $n2) { throw }
 
 "Or In Match"
-$n1 = Test-Query (query -Or (query Name Readme.txt), (query Name About.txt), (query Name LICENSE))
-$n2 = Test-Query (query Name -In Readme.txt, About.txt, LICENSE)
+$n1 = Test-Query (New-MdbcQuery -Or (New-MdbcQuery Name Readme.txt), (New-MdbcQuery Name About.txt), (New-MdbcQuery Name LICENSE))
+$n2 = Test-Query (New-MdbcQuery Name -In Readme.txt, About.txt, LICENSE)
 if ($n1 -ne $n2) { throw }
-$n2 = Test-Query (query Name -Match '^(?:Readme\.txt|About\.txt|LICENSE)$')
+$n2 = Test-Query (New-MdbcQuery Name -Match '^(?:Readme\.txt|About\.txt|LICENSE)$')
 if ($n1 -ne $n2) { throw }
 
 "Match, ignore case"
-$n1 = Test-Query (query Name -Match '^(?i:Readme\.txt|About\.txt|LICENSE)$')
-$n2 = Test-Query (query Name -Match (New-Object regex '^(?:Readme\.txt|About\.txt|LICENSE)$', IgnoreCase))
+$n1 = Test-Query (New-MdbcQuery Name -Match '^(?i:Readme\.txt|About\.txt|LICENSE)$')
+$n2 = Test-Query (New-MdbcQuery Name -Match (New-Object regex '^(?:Readme\.txt|About\.txt|LICENSE)$', IgnoreCase))
 if ($n1 -ne $n2) { throw }
 
 "Exists Mod Not"
-$MissingLength = Test-Query (query Length -Exists $false)
-$n1 = Test-Query (query Length -Mod 2, 0)
-$n2 = Test-Query (query Length -Not -Mod 2, 1)
+$MissingLength = Test-Query (New-MdbcQuery Length -Exists $false)
+$n1 = Test-Query (New-MdbcQuery Length -Mod 2, 0)
+$n2 = Test-Query (New-MdbcQuery Length -Not -Mod 2, 1)
 if ($MissingLength + $n1 -ne $n2) { throw }
-$n1 = Test-Query (query Length -Mod 2, 1)
-$n2 = Test-Query (query Length -Not -Mod 2, 0)
+$n1 = Test-Query (New-MdbcQuery Length -Mod 2, 1)
+$n2 = Test-Query (New-MdbcQuery Length -Not -Mod 2, 0)
 if ($MissingLength + $n1 -ne $n2) { throw }
 
 "Type"
-$n1 = Test-Query (query Length -Type Int64)
+$n1 = Test-Query (New-MdbcQuery Length -Type Int64)
 if ($n1 -ne $total - $MissingLength) { throw }
 
-if (0) {
-	"Where (VERY SLOW!)"
-	$n1 = Test-Query (query -Where 'this.Length == null')
-	if ($n1 -ne $MissingLength) { throw }
-}
+"Where (slow)"
+$n1 = Test-Query (New-MdbcQuery -Where 'this.Length == null')
+if ($n1 -ne $MissingLength) { throw }