+++ Tests/Test-Process.ps1
@@ -1,9 +1,9 @@
 
 <#
-.SYNOPSIS
+.Synopsis
 	Tests the Mdbc module with some process data
 
-.NOTES
+.Notes
 	TotalProcessorTime
 		It can be null.
 		Save as double, TimeSpan is not BSON type.
@@ -60,7 +60,7 @@ if ($n1 -ne $n2) { throw }
 "Count : $n1"
 
 ### Get by name
-$1 = Get-MdbcData $collection (query Name svchost) -Count
+$1 = Get-MdbcData $collection (New-MdbcQuery Name svchost) -Count
 $2 = $collection.Find([MongoDB.Driver.QueryDocument]@{ Name = 'svchost' }).Count()
 $3 = $collection.Find([MongoDB.Driver.Builders.Query]::EQ('Name', 'svchost')).Count()
 if ($1 -ne $2) { throw }
@@ -68,7 +68,7 @@ if ($1 -ne $3) { throw }
 "Find svchost : $1"
 
 ### Get by pattern/where
-$1 = Get-MdbcData $collection (query Name -Match '^svc|^mon') -Count
+$1 = Get-MdbcData $collection (New-MdbcQuery Name -Match '^svc|^mon') -Count
 $2 = $collection.Find([MongoDB.Driver.Builders.Query]::Matches('Name', '^svc|^mon')).Count()
 $3 = $collection.Find([MongoDB.Driver.QueryDocument]@{ '$where' = 'this.Name == "svchost" || this.Name == "mongod"' }).Count()
 if ($1 -ne $2) { throw }
@@ -76,8 +76,8 @@ if ($1 -ne $3) { throw }
 "Find regex : $$"
 "Find where : $$"
 
-Get-MdbcData $collection (query Name mongod) | Update-MdbcData $collection (update HandleCount -Increment 1)
-$document = Get-MdbcData $collection (query Name mongod)
+Get-MdbcData $collection (New-MdbcQuery Name mongod) | Update-MdbcData $collection (New-MdbcUpdate HandleCount -Increment 1)
+$document = Get-MdbcData $collection (New-MdbcQuery Name mongod)
 
 $document | Convert-MdbcData | Format-List | Out-String
 $document | Convert-MdbcJson