+++ Tests/Test-Update-PullAll.ps1
@@ -8,7 +8,7 @@ $$.Array = 1, 2, 3, 4, 1, 2, 3, 4
 $$ | Add-MdbcData $collection
 
 # update (PullAll) and get back
-$$ | Update-MdbcData $collection (update Array -PullAll 1, 2)
+$$ | Update-MdbcData $collection (New-MdbcUpdate Array -PullAll 1, 2)
 $$ = Get-MdbcData $collection
 
 # test: 1, 2 are removed; 3, 4 are there
@@ -19,7 +19,7 @@ if ($$.Array[2] -ne 3) { throw }
 if ($$.Array[3] -ne 4) { throw }
 
 # update (Pull) and get back
-$$ | Update-MdbcData $collection (update Array -Pull 3)
+$$ | Update-MdbcData $collection (New-MdbcUpdate Array -Pull 3)
 $$ = Get-MdbcData $collection
 
 # test: 3 is removed; 4 is there