+++ Tests/Test-Update-Rename.ps1
@@ -8,7 +8,7 @@ $$.Name1 = 42
 $$ | Add-MdbcData $collection
 
 # update (rename Name1 to Name2) and get back
-$$ | Update-MdbcData $collection (update Name1 -Rename Name2)
+$$ | Update-MdbcData $collection (New-MdbcUpdate Name1 -Rename Name2)
 $$ = Get-MdbcData $collection
 
 # test: Name2 gets 42, i.e. renamed